+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Changed
+* Made Error enum non-exhaustive
 
 ## [0.5.4] - 2020-09-11
 ### Added